ModernPythonDevelopment/Lab1/streamlit-page-master/pyproject.toml
[tool.black]
line-length = 80
exclude = '''
(
    \.DS_Store |
    \.githooks |
    \.html$ |
    \.idea |
    \.json$ |
    \.poetry |
    \.pyc$ |
    \.sql$ |
    \.vscode |
    \.venv |
    node_modules |
    \.yml$
)
'''

[tool.poetry]
name = "streamlit-page"
version = "0.1.0"
description = "Streamlit example project"
authors = ["Andrea Guzzo <andrea.guzzo92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^0.76.0"
pandas = "^1.2.2"
numpy = "^1.20.1"
plotly-express = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.5"
pylint = "^2.6.0"
flake8-bandit = "^2.1.2"
flake8-isort = "^4.0.0"
flake8-builtins = "^1.5.3"
black = "^20.8b1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
